Move CRUD methods to TransactionManager
