--- core/src/main/java/google/registry/schema/registrar/RegistrarDao.java
@@ -1,31 +1,0 @@
-// Copyright 2020 The Nomulus Authors. All Rights Reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package google.registry.schema.registrar;
-
-import google.registry.model.registrar.Registrar;
-import google.registry.schema.BasicDao;
-
-/** Data access object for {@link Registrar}. */
-public class RegistrarDao extends BasicDao {
-
-  private static final RegistrarDao INSTANCE = new RegistrarDao();
-
-  private RegistrarDao() {}
-
-  /** Returns a {@link RegistrarDao} instance. */
-  public static RegistrarDao registrarDao() {
-    return INSTANCE;
-  }
-}
