+++ core/src/test/java/google/registry/schema/registrar/RegistrarDaoTest.java
@@ -15,7 +15,7 @@
 package google.registry.schema.registrar;
 
 import static com.google.common.truth.Truth.assertThat;
-import static google.registry.schema.registrar.RegistrarDao.registrarDao;
+import static google.registry.persistence.transaction.TransactionManagerFactory.jpaTm;
 import static org.junit.Assert.assertThrows;
 
 import com.google.common.collect.ImmutableList;
@@ -64,32 +64,32 @@ public class RegistrarDaoTest extends EntityTestCase {
 
   @Test
   public void saveNew_worksSuccessfully() {
-    assertThat(registrarDao().checkExists(testRegistrar)).isFalse();
-    registrarDao().saveNew(testRegistrar);
-    assertThat(registrarDao().checkExists(testRegistrar)).isTrue();
+    assertThat(jpaTm().checkExists(testRegistrar)).isFalse();
+    jpaTm().saveNew(testRegistrar);
+    assertThat(jpaTm().checkExists(testRegistrar)).isTrue();
   }
 
   @Test
   public void update_worksSuccessfully() {
-    registrarDao().saveNew(testRegistrar);
-    Registrar persisted = registrarDao().load(registrarKey).get();
+    jpaTm().saveNew(testRegistrar);
+    Registrar persisted = jpaTm().load(registrarKey).get();
     assertThat(persisted.getRegistrarName()).isEqualTo("registrarName");
-    registrarDao().update(persisted.asBuilder().setRegistrarName("changedRegistrarName").build());
-    persisted = registrarDao().load(registrarKey).get();
+    jpaTm().update(persisted.asBuilder().setRegistrarName("changedRegistrarName").build());
+    persisted = jpaTm().load(registrarKey).get();
     assertThat(persisted.getRegistrarName()).isEqualTo("changedRegistrarName");
   }
 
   @Test
   public void update_throwsExceptionWhenEntityDoesNotExist() {
-    assertThat(registrarDao().checkExists(testRegistrar)).isFalse();
-    assertThrows(IllegalArgumentException.class, () -> registrarDao().update(testRegistrar));
+    assertThat(jpaTm().checkExists(testRegistrar)).isFalse();
+    assertThrows(IllegalArgumentException.class, () -> jpaTm().update(testRegistrar));
   }
 
   @Test
   public void load_worksSuccessfully() {
-    assertThat(registrarDao().checkExists(testRegistrar)).isFalse();
-    registrarDao().saveNew(testRegistrar);
-    Registrar persisted = registrarDao().load(registrarKey).get();
+    assertThat(jpaTm().checkExists(testRegistrar)).isFalse();
+    jpaTm().saveNew(testRegistrar);
+    Registrar persisted = jpaTm().load(registrarKey).get();
 
     assertThat(persisted.getClientId()).isEqualTo("registrarId");
     assertThat(persisted.getRegistrarName()).isEqualTo("registrarName");
